snapshots: Add create-url --uefi boolean flag
Differentiate between unset and false by using flags.Changed().

(Note the style in here leads to a lot of unecessary error checking --
and hence reference to string flag names. IMO it's nicer to create a
flags/vals struct and use the *Var* variants from pflags...)

(This is on top of https://github.com/vultr/govultr/pull/390)